+++ apps/desktop-wallet/src/api/apiDataHooks/address/types.ts
@@ -18,7 +18,7 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 
 import { AddressHash } from '@alephium/shared'
 
-import { SkipProp } from '@/api/apiDataHooks/types'
+import { SkipProp } from '@/api/apiDataHooks/apiDataHooksTypes'
 
 export interface UseFetchAddressProps extends SkipProp {
   addressHash: AddressHash
