+++ apps/desktop-wallet/src/api/apiDataHooks/address/useFetchAddressFts.ts
@@ -19,7 +19,7 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 import { AddressHash } from '@alephium/shared'
 
 import useFetchAddressTokensByType from '@/api/apiDataHooks/address/useFetchAddressTokensByType'
-import { SkipProp } from '@/api/apiDataHooks/types'
+import { SkipProp } from '@/api/apiDataHooks/apiDataHooksTypes'
 import useFetchSortedFts from '@/api/apiDataHooks/utils/useFetchSortedFts'
 
 interface UseAddressFTsProps extends SkipProp {
