+++ apps/desktop-wallet/src/api/apiDataHooks/apiDataHooksUtils.ts
