+++ apps/desktop-wallet/src/api/queries/addressQueries.ts
@@ -20,7 +20,7 @@ import { AddressHash, PAGINATION_PAGE_LIMIT, throttledClient } from '@alephium/s
 import { AddressTokenBalance } from '@alephium/web3/dist/src/api/api-explorer'
 import { queryOptions, skipToken } from '@tanstack/react-query'
 
-import { SkipProp } from '@/api/apiDataHooks/types'
+import { SkipProp } from '@/api/apiDataHooks/apiDataHooksTypes'
 import { AddressLatestTransactionHashQueryFnData } from '@/api/queries/transactionQueries'
 import queryClient from '@/api/queryClient'
 import { DisplayBalances, TokenDisplayBalances, TokenId } from '@/types/tokens'
