+++ apps/desktop-wallet/src/api/queries/priceQueries.ts
@@ -20,7 +20,7 @@ import { ONE_MINUTE_MS, throttledClient } from '@alephium/shared'
 import { ALPH } from '@alephium/token-list'
 import { queryOptions, skipToken } from '@tanstack/react-query'
 
-import { SkipProp } from '@/api/apiDataHooks/types'
+import { SkipProp } from '@/api/apiDataHooks/apiDataHooksTypes'
 import queryClient from '@/api/queryClient'
 
 interface TokensPriceQueryProps extends SkipProp {
